palaso.sfm.records: Make the field parser accept empty field values
The field parser always assumed a child for a marker even if it was an empty
string but fails if a marker is truly empty.  This went unnoticed because the
Windows line terminated usfm.sty file on linux generated a child with \r for
empty field values.  On windows the \r\n sequence was consumed as a line ending
and thus generating markers with no children at all.
Also update the history meta data in the source files.