Make tests pass on Py2 as well as Py3
All LDML results rely on knowing what the most common script is.
With a small test sample, the most common script might be a
common or inherited script, instead of specific script that would
would expect with a large data set. So more test data was added,
to make the distribution of scripts slightly more realistic.

In addition, the list of graphemes uses a Counter() object,
just like the collection of exemplars. The Counter() object
does not, AFAIK, provide the same order from the most_common()
method if the count of two items is the same.

So while these changes make the tests pass on Py2,
they should also make the tests more robust on Py3.