Adjust slightly the range of possible codepoints
sys.maxunicode has a value of 0x10FFFF, using this in range()
means that 0x10FFFE will be the highest codepoint that is looked
up (since the range starts with 0). However, 0x10FFFE causes an
the exception IndexError. Now that IndexError error is avoided,
the exception KeyError (to indicate not an assigned character)
can be used.