Default to uharfbuzz (uhb) for OpenType shaping
The older OpenType shaper (ot) is still available by
exporting CTR_UHB=0

For testing the different OpenType shapers (uhb and ot),
as well as Graphite with OpenType,
a new option (-n or --nocompare) has been added to output
all the shaping information. This way you don't have to modify
a font just to produce differences for cmptxtrender to find.